requirements.txt
requests~=2.31.0
ntplib~=0.4.0
dnspython~=2.5.0
inputimeout~=1.0.4
colorama~=0.4.6